Fix warnings in ntv2 on Linux
In ntv2bitfile fixed a sign mismatch comparison.
In ntv2nubaccess fixed an unused parameter.
In ntv2signalrouter fixed another instance of the catch (bad_alloc)
warning.

+++ ajalibraries/ajantv2/src/ntv2bitfile.cpp
@@ -104,7 +104,7 @@ string CNTV2Bitfile::ParseHeaderFromBuffer(const uint8_t* inBitfileBuffer, const
 	do
 	{
 		//	Preload bitfile header into mem
-		for (int i = 0; i < MAX_BITFILEHEADERSIZE - 1; i++) 
+		for (size_t i = 0; i < MAX_BITFILEHEADERSIZE - 1; i++)
 		{
 			if (i < inBufferSize) 
 			{
